+++ android/app/src/main/java/com/dragonguard/android/activity/MainActivity.kt
@@ -16,9 +16,7 @@ import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.Observer
 import com.bumptech.glide.Glide
 import com.dragonguard.android.R
-import com.dragonguard.android.UserDetailActivity
 import com.dragonguard.android.activity.compare.RepoChooseActivity
-import com.dragonguard.android.activity.menu.MenuActivity
 import com.dragonguard.android.activity.ranking.RankingsActivity
 import com.dragonguard.android.activity.search.SearchActivity
 import com.dragonguard.android.connect.NetworkCheck
@@ -26,7 +24,6 @@ import com.dragonguard.android.databinding.ActivityMainBinding
 import com.dragonguard.android.preferences.IdPreference
 import com.dragonguard.android.viewmodel.Viewmodel
 import kotlinx.coroutines.*
-import java.util.*
 
 /*
  사용자의 정보를 보여주고 검색, 랭킹등을
