+++ android/app/src/main/java/com/dragonguard/android/recycleradapter/RepositoryProfileAdapter.kt
@@ -9,7 +9,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.dragonguard.android.R
-import com.dragonguard.android.UserDetailActivity
+import com.dragonguard.android.activity.UserDetailActivity
 import com.dragonguard.android.activity.search.RepoContributorsActivity
 import com.dragonguard.android.model.search.RepoSearchResultModel
