+++ android/app/src/main/res/layout/activity_user_detail.xml
@@ -9,7 +9,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:weightSum="3.4"
-        tools:context=".UserDetailActivity">
+        tools:context=".activity.UserDetailActivity">
 
 
         <androidx.appcompat.widget.Toolbar
